Refactor code in client to fetch examples from the endpoint
